Эмбеддинг теперь не является дженериком. + скорректированы типы данных.п
+ правки в подключениях.

+++ src/textembedding/itextembedding.h
@@ -58,13 +58,13 @@ public:
     /// индексов символов для заданного текста
     /// \param startPos точка старта генерации
     /// \return 2д матрица с индексами символов текста
-    virtual Matrix2d<int> genTextIndices(int startPos) = 0;
+    virtual Matrix2d<double> genTextIndices(int startPos) = 0;
     ///
     /// \brief genTextBatch генерация партии (3д матрицы)
     /// представлений символов для заданного текста
     /// \param indices 2д матрица индексов символов текста
     /// \return 3д матрица с матем. представлениями символов текста
-    virtual Matrix3d<int> genTextBatch(Matrix2d<int> indices) = 0;
+    virtual Matrix3d<double> genTextBatch(Matrix2d<double> indices) = 0;
 
 public:
     virtual ~ITextEmbedding() {}
